models: Use smaller dropout ratios in large CNN.
Use smaller dropout ratios here as it seems that dropout may be the
reason behind the model's inability to train in a few cases. We suspect
that large dropout ratios lead to dying ReLUs [1]. However, we need to
further investigate this to be 100% sure. In our initial experiments
this solves the problem but decreases slightly (~0.8%) the model's
performance.

[1] https://cs231n.github.io/neural-networks-1/#actfun

Signed-off-by: Stefanos Koffas <koffas.stefanos@gmail.com>